Изменения после ревью

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/commerce/dto/DimensionDto.java
@@ -9,10 +9,10 @@ import lombok.*;
 @Builder
 public class DimensionDto {
     @NotNull
-    private double width;
+    private Float width;
     @NotNull
-    private double height;
+    private Float height;
     @NotNull
-    private double depth;
+    private Float depth;
 }
